move MapPrinter into MapPrinters folder, same for unit test
