+++ MarsRover.Tests/AppUI/Components/AppSectionObstaclesTests.cs
@@ -1,5 +1,6 @@
 ﻿using MarsRover.AppUI.Components;
 using MarsRover.AppUI.Helpers;
+using MarsRover.AppUI.MapPrinters;
 using MarsRover.AppUI.PositionStringFormat;
 using MarsRover.Controllers;
 using MarsRover.Models.Elementals;
